libobs-d3d11: Put d3d11 and d3d12 common functions together
