+++ libobs-d3d11/d3d11-shaderprocessor.hpp
@@ -28,10 +28,10 @@ struct ShaderProcessor {
 	gs_device_t *device;
 	ShaderParser parser;
 
-	void BuildInputLayout(vector<D3D11_INPUT_ELEMENT_DESC> &inputs);
-	void BuildParams(vector<gs_shader_param> &params);
-	void BuildSamplers(vector<unique_ptr<ShaderSampler>> &samplers);
-	void BuildString(string &outputString);
+	void BuildInputLayout(std::vector<D3D11_INPUT_ELEMENT_DESC> &inputs);
+	void BuildParams(std::vector<gs_shader_param> &params);
+	void BuildSamplers(std::vector<std::unique_ptr<ShaderSampler>> &samplers);
+	void BuildString(std::string &outputString);
 	void Process(const char *shader_string, const char *file);
 
 	inline ShaderProcessor(gs_device_t *device) : device(device) {}
